refactored the code into multiple sub-components
