+++ src/components/NavButtons.tsx
@@ -1,0 +1,10 @@
+
+function NavButtons(){
+    return(
+        <nav>
+          <button>Restaurants</button>
+          <button>Chefs</button>
+        </nav>
+    )
+}
+export default NavButtons
