+++ src/components/ShoppingcartToggle.tsx
@@ -1,0 +1,5 @@
+import { bagIcon } from "../assets/icons";
+function ShoppingcartToggle() {
+  return <img src={bagIcon} alt="bag-icon" />;
+}
+export default ShoppingcartToggle;
