+++ src/components/UserToggle.tsx
@@ -1,0 +1,5 @@
+import { userIcon } from "../assets/icons";
+function UserToggle() {
+  return <img src={userIcon} alt="user-icon" />;
+}
+export default UserToggle;
