+++ src/styles/LogoSection.scss
@@ -1,0 +1,10 @@
+div{
+    display: flex;
+    align-items: center;
+    span{
+        font-family: Helvetica;
+        font-weight: 200;
+        font-size: 27px;
+        letter-spacing: 1.35px;
+    }
+}
